Work-around snakemake issues with checkpoints
Snakemake apparently has an issue where it can get into a state
that always reruns jobs if you have two checkpoints. See:
https://github.com/snakemake/snakemake/issues/2175
To work around this, we need to force the checkpoints to be run before
other jobs. Sometimes this seems to require adding the checkpoint output file
to the downstream jobs' inputs and sometimes it seems to require using the
`checkpoint` object itself, or both. Some of these changes may be unnecessary
as they are presumably force through relationships with other rules but including
them does no harm.